requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
crispy-bootstrap4==2025.6
cron-descriptor==1.4.5
Django==5.2.3
django-celery-beat==2.8.1
django-crispy-forms==2.4
django-jazzmin==3.0.1
django-modeltranslation==0.19.15
django-rosetta==0.10.2
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
flower==2.0.1
humanize==4.12.3
idna==3.10
inflection==0.5.1
kombu==5.5.4
packaging==25.0
pillow==11.2.1
polib==1.2.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tornado==6.5.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
